moved all methods from BlueshiftAttrDevice to BlueshiftAttrApp to avoid confusions
